refactor: move create game instances in separates files
