+++ src/webSocket_server/db/playersOnline.ts
@@ -1,0 +1,5 @@
+import OnlinePlayers from "../services/onlinePlayersService";
+
+const playersOnline = new OnlinePlayers();
+
+export default playersOnline;
